Start a major refactoring of Sequence and related classes
- Sequence was mainly a container of EIT data. This responsibility has
been moved to EITData.
- Subclasses of EITData (previously subclasses of Sequence) for
different vendors are in separate modules.
- The responsibility of Frameset has moved to EITDataVariant.
- Visualisation tools have been moved to a separate module `plotting`.
- The Vendor class has been moved to it's own module.
- `merge()` methods are now called `concatenate()`.

This is the start of the road for this refactoring. Up to this point,
there is very limited functionality and testing.

What works*:
- DraegerEITData.from_path() works.
- EITData.from_path(..., vendor="draeger") works.
- DraegerEITData.concatenate() works. This includes
VariantCollection.concatenate() and EITVariant.concatenate().
- DraegerEITData.check_equivalence() is working.

*: Untested!

Interface:
a = DraegerEITData.from_path("test_data/a.bin")
b = DraegerEITData.from_path("test_data/b.bin")
c = a + b
c.variants['raw'].pixel_impedance
