+++ eitprocessing/helper.py
@@ -1,0 +1,2 @@
+class NotEquivalent(Exception):
+    """Raised when two objects are not equivalent."""
